partial 0.9.4, support for file streams and script

+++ divconq.core/src/main/java/divconq/lang/WrappedFuncCallback.java
@@ -84,7 +84,7 @@ abstract public class WrappedFuncCallback<T> extends FuncCallback<T> {
 	}
 	
 	@Override
-	public long getAmountCompleted() {
+	public int getAmountCompleted() {
 		return this.or.getAmountCompleted();
 	}
